// models/linkType.ts
export enum LinkType {
  Any = 'ZZ',
  Whatsapp = 'WA',
  Line = 'LN',
  Telegram = 'TG',
  Tokopedia = 'TP',
  Shopee = 'SP',
  Lazada = 'LZ',
  GoFood = 'GO',
  GrabFood = 'GR',
  TravelokaEats = 'TE',
  ShopeeFood = 'SF',
  Instagram = 'IG',
  Facebook = 'FB',
  Twitter = 'TW',
  TikTok = 'TT',
  Youtube = 'YT',
  SO_Web = 'SO_Web',
}